+++ HISTORY.md
@@ -1,5 +1,11 @@
 # History
 
+## 0.3.0 (2025-09-13)
+
+* Fix so local caching via TOML works.
+* `get_secret()`: Add `force_refresh` and `raw` parameters
+* Attempt to `JSON.loads` secret string from AWS Secrets Manager by default.
+
 ## 0.2.0 (2025-09-13)
 
 * Add core caching logic.